isolate logic in hooks
